+++ frontend/src/Chat/components/ChatRoom/BanButton.tsx
@@ -1,0 +1,75 @@
+import { Button } from 'react-bootstrap'
+import { BannedIcon } from '../utils/BannedIcon'
+import { updateChatRoomMembersRequest } from '../utils/requestUtils'
+
+const BanMemberButton = (props: {
+  room: ChatRoom
+  member: User
+  updateMemberList: () => void
+}): JSX.Element => {
+  return (
+    <Button
+      variant="outline-danger"
+      onClick={() => {
+        const requestData: ChatRoomMember = {
+          chatRoomId: props.room.id,
+          userId: props.member.id,
+          isBanned: true,
+          isAdministrator: false
+        }
+        updateChatRoomMembersRequest(requestData, props.updateMemberList)
+      }}
+    >
+      Ban
+    </Button>
+  )
+}
+
+const BanOFFMemberButton = (props: {
+  room: ChatRoom
+  member: User
+  updateMemberList: () => void
+}): JSX.Element => {
+  return (
+    <Button
+      variant="outline-info"
+      onClick={() => {
+        const requestData: ChatRoomMember = {
+          chatRoomId: props.room.id,
+          userId: props.member.id,
+          isBanned: false,
+          isAdministrator: false
+        }
+        updateChatRoomMembersRequest(requestData, props.updateMemberList)
+      }}
+    >
+      Ban 解除
+    </Button>
+  )
+}
+
+export const BanButton = (props: {
+  room: ChatRoom
+  member: User
+  chatRoomMemberList: ChatRoomMember[]
+  updateMemberList: () => void
+}): JSX.Element => {
+  const isOwner: boolean = props.room.created_by_user_id === props.member.id
+  if (isOwner) return <></>
+
+  const isBanned = props.chatRoomMemberList.find(
+    (item) =>
+      item.userId === props.member.id && item.chatRoomId === props.room.id
+  )?.isBanned
+  if (isBanned === undefined) return <></>
+  return isBanned ? (
+    <>
+      <BannedIcon isBanned={isBanned} />
+      <BanOFFMemberButton {...props} member={props.member} />
+    </>
+  ) : (
+    <>
+      <BanMemberButton {...props} member={props.member} />
+    </>
+  )
+}
