+++ frontend/src/Chat/components/ChatRoom/UserListDisplay.tsx
@@ -1,156 +1,9 @@
 import { type ReactElement } from 'react'
 import { Button } from 'react-bootstrap'
-import { AdminIcon } from '../utils/AdminIcon'
-import { BannedIcon } from '../utils/BannedIcon'
 import { OwnerIcon } from '../utils/OwnerIcon'
-import {
-  deleteChatRoomMembersRequest,
-  updateChatRoomMembersRequest
-} from './requestUtils'
-
-const BanMemberButton = (props: {
-  room: ChatRoom
-  member: User
-  updateMemberList: () => void
-}): JSX.Element => {
-  return (
-    <Button
-      variant="outline-danger"
-      onClick={() => {
-        const requestData: ChatRoomMember = {
-          chatRoomId: props.room.id,
-          userId: props.member.id,
-          isBanned: true,
-          isAdministrator: false
-        }
-        updateChatRoomMembersRequest(requestData, props.updateMemberList)
-      }}
-    >
-      Ban
-    </Button>
-  )
-}
-
-const BanOFFMemberButton = (props: {
-  room: ChatRoom
-  member: User
-  updateMemberList: () => void
-}): JSX.Element => {
-  return (
-    <Button
-      variant="outline-info"
-      onClick={() => {
-        const requestData: ChatRoomMember = {
-          chatRoomId: props.room.id,
-          userId: props.member.id,
-          isBanned: false,
-          isAdministrator: false
-        }
-        updateChatRoomMembersRequest(requestData, props.updateMemberList)
-      }}
-    >
-      Ban 解除
-    </Button>
-  )
-}
-
-const BanButton = (props: {
-  room: ChatRoom
-  member: User
-  chatRoomMemberList: ChatRoomMember[]
-  updateMemberList: () => void
-}): JSX.Element => {
-  const isOwner: boolean = props.room.created_by_user_id === props.member.id
-  if (isOwner) return <></>
-
-  const isBanned = props.chatRoomMemberList.find(
-    (item) =>
-      item.userId === props.member.id && item.chatRoomId === props.room.id
-  )?.isBanned
-  if (isBanned === undefined) return <></>
-  return isBanned ? (
-    <>
-      <BannedIcon isBanned={isBanned} />
-      <BanOFFMemberButton {...props} member={props.member} />
-    </>
-  ) : (
-    <>
-      <BanMemberButton {...props} member={props.member} />
-    </>
-  )
-}
-
-const AdminMemberButton = (props: {
-  room: ChatRoom
-  member: User
-  updateMemberList: () => void
-}): JSX.Element => {
-  return (
-    <Button
-      variant="outline-info"
-      onClick={() => {
-        const requestData: ChatRoomMember = {
-          chatRoomId: props.room.id,
-          userId: props.member.id,
-          isBanned: false,
-          isAdministrator: true
-        }
-        updateChatRoomMembersRequest(requestData, props.updateMemberList)
-      }}
-    >
-      Admin
-    </Button>
-  )
-}
-
-const AdminOFFMemberButton = (props: {
-  room: ChatRoom
-  member: User
-  updateMemberList: () => void
-}): JSX.Element => {
-  return (
-    <Button
-      variant="outline-danger"
-      onClick={() => {
-        const requestData: ChatRoomMember = {
-          chatRoomId: props.room.id,
-          userId: props.member.id,
-          isBanned: false,
-          isAdministrator: false
-        }
-        updateChatRoomMembersRequest(requestData, props.updateMemberList)
-      }}
-    >
-      Admin 解除
-    </Button>
-  )
-}
-
-const AdminButton = (props: {
-  room: ChatRoom
-  member: User
-  chatRoomMemberList: ChatRoomMember[]
-  updateMemberList: () => void
-}): JSX.Element => {
-  const isOwner: boolean = props.room.created_by_user_id === props.member.id
-  if (isOwner) return <></>
-
-  const isAdmin = props.chatRoomMemberList.find(
-    (item) =>
-      item.userId === props.member.id && item.chatRoomId === props.room.id
-  )?.isAdministrator
-  if (isAdmin === undefined) return <></>
-  return isAdmin ? (
-    <>
-      <AdminIcon isAdmin={isAdmin} />
-      <AdminOFFMemberButton {...props} member={props.member} />
-    </>
-  ) : (
-    <>
-      <AdminMemberButton {...props} member={props.member} />
-    </>
-  )
-}
+import { AdminButton } from './AdminButton'
+import { BanButton } from './BanButton'
+import { deleteChatRoomMembersRequest } from '../utils/requestUtils'
 
 const DeleteMemberButton = (props: {
   room: ChatRoom
